♻️ REFACTOR: 유저 관련 파일 폴더 위치 변경
REFACTOR domain/user/*
    - 유저 관련 파일 폴더 위치 변경
