+++ src/main/java/com/sparta/fooddeliveryapp/domain/user/repository/UserRepository.java
@@ -1,6 +1,6 @@
-package com.sparta.fooddeliveryapp.domain.review.repository;
+package com.sparta.fooddeliveryapp.domain.user.repository;
 
-import com.sparta.fooddeliveryapp.domain.review.entity.User;
+import com.sparta.fooddeliveryapp.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
